+++ src/account/handle.rs
@@ -111,7 +111,10 @@ impl AccountHandle {
         let foundry_output_id = self.client.foundry_output_id(foundry_id).await?;
         let output_response = self.client.get_output(&foundry_output_id).await?;
 
-        Ok(Output::try_from(&output_response.output)?)
+        Ok(Output::try_from_dto(
+            &output_response.output,
+            self.client.get_token_supply()?,
+        )?)
     }
 
     /// Get the [`Transaction`] of a transaction stored in the account